Packaging: fix installing in a directory with GNU Make
Packaging usually involves installing a file into a directory, which is
passed to make through the DESTDIR variable.

This commit adds ${DESTDIR} as a prefix to the installation directories,
and also ensures the directories exist before copying the files over.

It is worth noting the BSD makefile already does that.